+++ Apsitvarkom.Api/appsettings.Staging.json
@@ -5,5 +5,8 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "Geocoding" : {
+    "Url": "https://maps.googleapis.com/maps/api/geocode/json/"
   }
 }
